FIX: for `git_compare_dates`, modified the Git repo check
before this commit, the Git repo check would fail unless the script was run from the root of the Git repo.

now, the script can be run from any and all subdirectories within a Git repo, and the script will correctly identify that a parent directory is a Git repo.