FIX: `git_package_manager` now comments on repos where HEAD didn't move
before this commit, running `git fetch` and `git pull` within a repo for which `HEAD` is not at the tip of the branch would result in an output saying the repo updated successfully. running `git_package_manager` again would result in the same output, which doesn't make any sense.

now, `git_package_manager` correctly identifies repos for which `HEAD` is not at the tip of the branch, and outputs a message saying that the repo was updated, but `HEAD` wasn't moved.