FIX: `ansible_encrypt` now correctly errors when passed an invalid salt
this commit updates `ansible_encrypt` so that it correctly errors when an invalid salt is provided.

before this commit, `$1` would have to be `encrypt` for the script to error. but, `$1` is never `encrypt` at this point, because `shift` has not yet been called.

with this commit, the script will correctly identify an "invalid salt" as either of the following:
- when a salt is not provided (`-s` would have to be the last option, with no argument passed)
- when the argument for `-s` is another option (any string beginning with `-`)